requirements.txt
inplace_abn==1.1.0
matplotlib==3.5.2
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.1
pycocotools==2.0
PyYAML==6.0
randaugment==1.0.2
